Making some basic baseline checks
